requirements.txt
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.29.0
livekit-api==1.0.7
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
